refactor(backend): Modularize Rust thumbnail engine and commands
REFACTORING:
Thumbnail Engine Modularization:
- Split monolithic thumbnail_engine.rs into focused modules
- Created thumbnail_engine/cache.rs for caching logic
- Created thumbnail_engine/queue.rs for queue management
- Created thumbnail_engine/retry.rs for retry mechanisms
- Created thumbnail_engine/types.rs for shared types
- Reduced main thumbnail_engine.rs by ~1300 lines

Commands Organization:
- Extracted thumbnail commands to commands/thumbnail.rs
- Updated commands/mod.rs to include new module
- Reduced lib.rs by ~1100 lines

Module Structure:
src-tauri/src/
├── lib.rs (simplified)
├── commands/
│   ├── mod.rs
│   └── thumbnail.rs (new)
└── thumbnail_engine/
    ├── mod.rs (main)
    ├── cache.rs (new)
    ├── queue.rs (new)
    ├── retry.rs (new)
    └── types.rs (new)

BENEFITS:
✅ Reduced lib.rs from ~1100 lines
✅ Reduced thumbnail_engine.rs by ~1300 lines
✅ Clear module boundaries
✅ Easier to maintain and test
✅ Better code organization
✅ Improved compilation times

FILES CHANGED:
- Modified: lib.rs, thumbnail_engine.rs, commands/mod.rs
- Created: commands/thumbnail.rs, thumbnail_engine/{cache,queue,retry,types}.rs
- Net reduction: ~2400 lines in main files
