+++ src-tauri/src/commands/mod.rs
@@ -1,7 +1,9 @@
 pub mod media;
 pub mod project;
 pub mod export;
+pub mod thumbnail;
 
 pub use media::*;
 pub use project::*;
 pub use export::*;
+pub use thumbnail::*;
